requirements.txt
Flask==3.0.3
python-dateutil==2.8.2
emoji==2.10.0 